Refactor: Service Inteface 생성
