+++ Pick/src/main/java/com/nob/pick/report/query/controller/ReportController.java
@@ -4,12 +4,9 @@ import com.nob.pick.report.query.dto.ReportDTO;
 import com.nob.pick.report.query.service.ReportService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.nio.charset.Charset;
 import java.util.List;
 
 // 조회 관련 API 요청 처리 (DTO 반환)
